re-writing main ppu loop.
The next place to check for bugs is here, and the code here was extrememly messy so I decided to rewrite it. I considered how to make it more organized, and it seems best to check conditions for each event seperately instead of trying to lump them together by cycle or scanline.